Bring back cypher projection for plugin
and use .native for normal projection

# Conflicts:
#	src/graphdatascience/procedure_surface/cypher/catalog/catalog_cypher_endpoints.py
