+++ src/graphdatascience/procedure_surface/cypher/catalog/__init__.py
@@ -1,13 +1,19 @@
 from graphdatascience.procedure_surface.cypher.catalog.catalog_cypher_endpoints import (
     CatalogCypherEndpoints,
-    GraphNativeProjectEndpoints,
+)
+from graphdatascience.procedure_surface.cypher.catalog.projection_cypher_endpoints import (
+    GraphCypherProjectResult,
     GraphProjectResult,
+    GraphWithCypherProjectResult,
     GraphWithProjectResult,
+    ProjectCypherEndpoints,
 )
 
 __all__ = [
     "CatalogCypherEndpoints",
-    "GraphNativeProjectEndpoints",
+    "GraphCypherProjectResult",
     "GraphProjectResult",
+    "GraphWithCypherProjectResult",
     "GraphWithProjectResult",
+    "ProjectCypherEndpoints",
 ]
