+++ tests/integration/procedure_surface/cypher/model/test_model_catalog_cypher_endpoints.py
@@ -27,7 +27,7 @@ def sample_graph(gds: GraphDataScience) -> Generator[str, None, None]:
         (c)-[:REL]->(a)
         """
     )
-    G, _ = gds.graph.project("model_catalog_cypher_g", {"Node": {"properties": ["age"]}}, "REL")
+    G, _ = gds.graph.project.native("model_catalog_cypher_g", {"Node": {"properties": ["age"]}}, "REL")
 
     yield G.name()
